[REFACTOR] change main file name (#4)
